LayoutTests/platform/glib/accessibility/element-line-rects-and-text-expected.txt

This is some testing content. A link here. MOre text.
z
Another line of text.

Jack and Jill went up the hill to fetch a pail of water. Jack fell down...
 abcdefghijklmnopqrstuvwxyz0123456789 abcdefghijklmnopqrstuvwxyz0123456789 abcdefghijklmnopqrstuvwxyz0123456789-abcdefghijklmnopqrstuvwxyz0123456789-abcdefghijklmnopqrstuvwxyz0123456789

This tests that lineRectsAndText returns the correct values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
